+++ packages/schema/_src/_api/string.ts
@@ -33,104 +33,138 @@ export const NonEmptyString = extendWithUtils(S.nonEmptyString)
 export type NonEmptyString = S.NonEmptyString
 
 /**
- * A string that is at least 1 character long and a maximum of 255.
+ * A string that is at least 1 character long and a maximum of 50.
  */
-export interface ReasonableStringBrand extends LongStringBrand {
-  readonly ReasonableString: unique symbol
+export interface NonEmptyString50Brand extends NonEmptyString255Brand {
+  readonly NonEmptyString50: unique symbol
 }
 
 /**
- * A string that is at least 1 character long and a maximum of 255.
+ * A string that is at least 1 character long and a maximum of 50.
  */
-export type ReasonableString = string & ReasonableStringBrand
+export type NonEmptyString50 = string & NonEmptyString50Brand
 
 /**
- * A string that is at least 1 character long and a maximum of 255.
+ * A string that is at least 1 character long and a maximum of 50.
  */
-export const reasonableStringFromString = pipe(
-  makeConstrainedFromString<ReasonableString>(1, 256 - 1),
+export const nonEmptyString50FromString = pipe(
+  makeConstrainedFromString<NonEmptyString50>(1, 256 - 1),
   S.arbitrary((FC) =>
     FC
       .lorem({ mode: "words", maxCount: 2 })
       .filter((x) => x.length < 256 - 1 && x.length > 0)
-      .map((x) => x as ReasonableString)
+      .map((x) => x as NonEmptyString50)
   ),
   // arbitrary removes brand benefit
-  S.brand<ReasonableString>()
+  S.brand<NonEmptyString50>()
+)
+
+/**
+ * A string that is at least 1 character long and a maximum of 50.
+ */
+export const NonEmptyString50 = extendWithUtils(
+  pipe(S.string[">>>"](nonEmptyString50FromString), S.brand<NonEmptyString50>())
 )
 
 /**
  * A string that is at least 1 character long and a maximum of 255.
  */
-export const ReasonableString = extendWithUtils(
-  pipe(S.string[">>>"](reasonableStringFromString), S.brand<ReasonableString>())
+export interface NonEmptyString255Brand extends NonEmptyString2kBrand {
+  readonly NonEmptyString255: unique symbol
+}
+
+/**
+ * A string that is at least 1 character long and a maximum of 255.
+ */
+export type NonEmptyString255 = string & NonEmptyString255Brand
+
+/**
+ * A string that is at least 1 character long and a maximum of 255.
+ */
+export const nonEmptyString255FromString = pipe(
+  makeConstrainedFromString<NonEmptyString255>(1, 256 - 1),
+  S.arbitrary((FC) =>
+    FC
+      .lorem({ mode: "words", maxCount: 2 })
+      .filter((x) => x.length < 256 - 1 && x.length > 0)
+      .map((x) => x as NonEmptyString255)
+  ),
+  // arbitrary removes brand benefit
+  S.brand<NonEmptyString255>()
+)
+
+/**
+ * A string that is at least 1 character long and a maximum of 255.
+ */
+export const NonEmptyString255 = extendWithUtils(
+  pipe(S.string[">>>"](nonEmptyString255FromString), S.brand<NonEmptyString255>())
 )
 
 /**
  * A string that is at least 1 character long and a maximum of 2047.
  */
-export interface LongStringBrand extends TextStringBrand {
-  readonly LongString: unique symbol
+export interface NonEmptyString2kBrand extends NonEmptyString64kBrand {
+  readonly NonEmptyString2k: unique symbol
 }
 
 /**
  * A string that is at least 1 character long and a maximum of 2047.
  */
-export type LongString = string & LongStringBrand
+export type NonEmptyString2k = string & NonEmptyString2kBrand
 
 /**
  * A string that is at least 1 character long and a maximum of 2047.
  */
-export const longStringFromString = pipe(
-  makeConstrainedFromString<LongString>(1, 2048 - 1),
+export const nonEmptyString2kFromString = pipe(
+  makeConstrainedFromString<NonEmptyString2k>(1, 2048 - 1),
   S.arbitrary((FC) =>
     FC
       .lorem({ mode: "words", maxCount: 25 })
       .filter((x) => x.length < 2048 - 1 && x.length > 0)
-      .map((x) => x as LongString)
+      .map((x) => x as NonEmptyString2k)
   ),
   // arbitrary removes brand benefit
-  S.brand<LongString>()
+  S.brand<NonEmptyString2k>()
 )
 
 /**
  * A string that is at least 1 character long and a maximum of 2047.
  */
-export const LongString = extendWithUtils(
-  pipe(S.string[">>>"](longStringFromString), S.brand<LongString>())
+export const NonEmptyString2k = extendWithUtils(
+  pipe(S.string[">>>"](nonEmptyString2kFromString), S.brand<NonEmptyString2k>())
 )
 
 /**
  * A string that is at least 1 character long and a maximum of 64kb.
  */
-export interface TextStringBrand extends NonEmptyBrand {
-  readonly TextString: unique symbol
+export interface NonEmptyString64kBrand extends NonEmptyBrand {
+  readonly NonEmptyString64k: unique symbol
 }
 
 /**
  * A string that is at least 1 character long and a maximum of 64kb.
  */
-export type TextString = string & TextStringBrand
+export type NonEmptyString64k = string & NonEmptyString64kBrand
 
 // TODO: compose arbitraries?
 /**
  * A string that is at least 1 character long and a maximum of 64kb.
  */
-export const textStringFromString = pipe(
-  makeConstrainedFromString<TextString>(1, 64 * 1024),
+export const nonEmptyString64kFromString = pipe(
+  makeConstrainedFromString<NonEmptyString64k>(1, 64 * 1024),
   S.arbitrary((FC) =>
     FC
       .lorem({ mode: "sentences", maxCount: 25 })
       .filter((x) => x.length < 64 * 1024 && x.length > 0)
-      .map((x) => x as TextString)
+      .map((x) => x as NonEmptyString64k)
   ),
   // arbitrary removes brand benefit
-  S.brand<TextString>()
+  S.brand<NonEmptyString64k>()
 )
 
 /**
  * A string that is at least 1 character long and a maximum of 64kb.
  */
-export const TextString = extendWithUtils(
-  pipe(S.string[">>>"](textStringFromString), S.brand<TextString>())
+export const NonEmptyString64k = extendWithUtils(
+  pipe(S.string[">>>"](nonEmptyString64kFromString), S.brand<NonEmptyString64k>())
 )