+++ packages/fluent-extensions/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @effect-app/fluent-extensions
 
+## 0.83.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @effect-app/schema@0.76.0
+
 ## 0.83.0
 
 ### Minor Changes